M6 Task 1 review fixes: t12_quarantined_rows is BIGINT, and a falsifiable finite-log test
Finding 1: DuckDB promotes sum() over a BIGINT column to HUGEINT, so the
column shipped as HUGEINT against the brief's BIGINT spec -- silent today,
but Task 6 reads it through @duckdb/node-api into TypeScript. Cast back to
::BIGINT after the FILTER clause (a cast between sum() and FILTER is a
parser error) and added a test asserting the type itself via
information_schema.columns, not just presence/sign.

Finding 2: test_health_score_is_always_finite could never go red on any
data this project has -- no route in the fixture or the real 2026-04
warehouse has gauge_t12 = 0 (min measured 0.958). Replaced with a test
against a tiny in-test relation that feeds the guarded and unguarded
ln(nullif(...)) expressions directly. That also surfaced that DuckDB's
ln(0) does not return -inf as the design doc's A.7 assumed -- it raises
'Out of Range Error: cannot take logarithm of zero', a hard runtime error
that would abort the whole mart build rather than merely mis-sort a row.
The nullif() guard's real job is keeping ln() from ever seeing a literal
zero. Flagged for a docs/data/model.md correction in a later task rather
than fixed here, since this task is scoped to SQL + tests only.

+++ sql/02_marts/200_mart_route_health.sql
@@ -40,7 +40,12 @@ agg AS (
         sum(r.passengers)           FILTER (WHERE r.year_month BETWEEN w.t12_start_month AND w.t12_end_month) AS t12_passengers,
         sum(r.departures_performed) FILTER (WHERE r.year_month BETWEEN w.t12_start_month AND w.t12_end_month) AS t12_departures_performed,
         sum(r.departures_scheduled) FILTER (WHERE r.year_month BETWEEN w.t12_start_month AND w.t12_end_month) AS t12_departures_scheduled,
-        sum(r.quarantined_rows)     FILTER (WHERE r.year_month BETWEEN w.t12_start_month AND w.t12_end_month) AS t12_quarantined_rows,
+        -- ::BIGINT after the FILTER (a cast between sum() and FILTER is a parser error):
+        -- DuckDB promotes sum() over a BIGINT column to HUGEINT. Left unguarded the column
+        -- ships as HUGEINT, not the BIGINT the brief specifies -- silent here, but Task 6
+        -- reads this through @duckdb/node-api into TypeScript, and this repo has a
+        -- documented history of DuckDB runtime types surfacing differently than expected.
+        sum(r.quarantined_rows) FILTER (WHERE r.year_month BETWEEN w.t12_start_month AND w.t12_end_month)::BIGINT AS t12_quarantined_rows,
 
         count(DISTINCT r.year_month) FILTER (
             WHERE r.year_month BETWEEN w.p12_start_month AND w.p12_end_month) AS p12_months_present,
